Implement some REST methods for Channel
Several REST methods for the Channel are implemented.
However, due to lack of support for 'Permission' and
'Message' at this stage, the corresponding methods
for messaging and permission handling will throw a
'notImplemented' error. This serves as a placeholder
indicating that these features are yet to be implemented.

These initial REST methods lay the foundation for future
development and expansion of the functionality.